fix(db): fold musehub_issue_milestones into 0001 migration; add MusehubIssueMilestone ORM to musehub_models
- Deleted 0002_milestones.py (violates one-migration policy; all schema
  changes during development belong in 0001_consolidated_schema.py)
- Deleted musehub_milestone_models.py (duplicate MusehubMilestone class
  conflicts with musehub_models.py; incompatible schema mismatch)
- Added musehub_issue_milestones table to 0001 upgrade/downgrade
- Added MusehubIssueMilestone ORM class to musehub_models.py
- Added issue_milestones relationship on existing MusehubMilestone
- Added Index import to musehub_models.py
